chore: moved write baseline
